WIP Allow for building packaging changes
Some of this could be pushed into the patch it depends on instead,
but with this patch we can build the following config:

```
artg_change_list:
  - project: openstack/ironic
    branch: master
    packaging_repo: https://review.rdoproject.org/r/openstack/ironic-distgit
    packaging_refspec: refs/changes/57/1157/6
  - project: openstack/puppet-nova
    branch: master
    code_repo: https://git.openstack.org/openstack/puppet-nova
    code_refspec: refs/changes/41/316241/8
  - project: openstack/cinder
    branch: master
    code_repo: https://git.openstack.org/openstack/cinder
    code_refspec: refs/changes/44/318744/3
    packaging_repo: http://review.rdoproject.org/r/openstack/cinder-distgit
    packaging_refspec: refs/changes/48/1148/4
```

I also hacked out the DLRN step where it installs the packages, because
puppet-nova was failing otherwise. (while installing it worked fine on
an actual undercloud) This could be moved to another patch.