Update patch set 4
Patch Set 4: Code-Review-1

(4 comments)

I would be willing to do a follow up to transform the python scripts into modules if you like the idea but not the idea of making the effort :).

The -1 is about jenkins_changes.py.

I also agree that doing the transform from host->code_repo or packaging_repo can be done in build.yml in my follow up in a way that allows code changes to depend on packaging changes and vice versa.

Patch-set: 4
Label: Code-Review=-1