Update patch set 4
Patch Set 4: Code-Review+1

Looks good to me, I think you can start bringing it out of WIP stage.

I rebased it on a "small fixes" patch to not be able to submit that workaround for the rpm install test quicker.

I'm running a local test with this (only code changes) and it passed the package building already.

Patch-set: 4
Label: Code-Review=+1