Update patch set 6
Patch Set 6: Code-Review+1

(1 comment)

This looks good to me. Tracking down the variables generated by Zuul was not straightforward, but it seems to check out.

Are the new variables in artg_change_list going to come in a separate review?

Unrelatedly, the build failures seem to be caused by packaging problems related to conflicting file ownership.

Patch-set: 6
Label: Code-Review=+1